Module Naboris

val listen : ?⁠inetAddr:Unix.inet_addr -> int -> 'sessionData ServerConfig.t -> 'a Lwt.t * 'a Lwt.u

Will start an http server listening at inetAddr on port int with ServerConfig.t('sessionData).

val listenAndWaitForever : ?⁠inetAddr:Unix.inet_addr -> int -> 'sessionData ServerConfig.t -> 'a Lwt.t

Same as listen but will specifically throw away the Lwt.u('a) and never resolve the promise. Keeping the server alive until the process is killed.

module Cookie : sig ... end

module Method : sig ... end

module Router : sig ... end

module Query : sig ... end

module MimeTypes : sig ... end

module Req : sig ... end

module Res : sig ... end

module Route : sig ... end

module ServerConfig : sig ... end

module Session : sig ... end

module SessionManager : sig ... end