Module Naboris.SessionManager

Less commonly used.

val startSession : 'sessionData Req.t -> Res.t -> 'sessionData -> 'sessionData Req.t * Res.t * string

Intended for internal use.

Create a new sessino with 'sessionData and add cookie headers to Res.t. Returns newly created session id string.

val resumeSession : 'sessionData ServerConfig.t -> 'sessionData Req.t -> 'sessionData Req.t Lwt.t

Intended for internal use.

Applies sessionGetter from config to request which uses the session id from the request cookies. Returns promise of a new request with session data available if it was found.