Module Naboris__.SessionManager

val startSession : 'sessionData Naboris.Req.t -> Naboris.Res.t -> 'sessionData -> 'sessionData Naboris.Req.t * Naboris.Res.t * string

Create a new session with 'sessionData and add cookie headers to Res.t. Returns newly created session id string.

val removeSession : 'sessionData Naboris.Req.t -> Naboris.Res.t -> Naboris.Res.t

Sets headers on `Res.t` to expire the session.

val resumeSession : 'sessionData Naboris.ServerConfig.t -> 'sessionData Naboris.Req.t -> 'sessionData Naboris.Req.t Lwt.t

Intended for internal use.

Applies mapSession from config to request which uses the session id from the request cookies. Returns promise of a new request with session data available if it was found.