Module Naboris__SessionManager

val startSession : 'sessionData Naboris.Req.t -> Naboris.Res.t -> 'sessionData -> 'sessionData Naboris.Req.t * Naboris.Res.t * string

* Intended for internal use. * * Create a new sessino with `'sessionData` and add cookie headers to `Res.t`. Returns newly created session id `string`.

val resumeSession : 'sessionData Naboris.ServerConfig.t -> 'sessionData Naboris.Req.t -> 'sessionData Naboris.Req.t Lwt.t

* Intended for internal use. * * Applies `sessionGetter` from config to request which uses the session id from the request cookies. * Returns promise of a new request with session data available if it was found.