Module Naboris

val listen : ?⁠inetAddr:Unix.inet_addr -> int -> 'sessionData ServerConfig.t -> 'a Lwt.t * 'a Lwt.u

Will start an http server listening at inetAddr on port int with ServerConfig.t('sessionData).

val listenAndWaitForever : ?⁠inetAddr:Unix.inet_addr -> int -> 'sessionData ServerConfig.t -> 'a Lwt.t

Same as listen but will specifically throw away the Lwt.u('a) and never resolve the promise. Keeping the server alive until the process is killed.

module ErrorHandler : sig ... end

Module with error handler types.

module Req : sig ... end

Module for working with incoming requests.

module Res : sig ... end

Module for creating and sending responses.

module Route : sig ... end

Module to extract routing data.

module ServerConfig : sig ... end

Module for configuring the naboris http server.

module Session : sig ... end

Module for working with sessions and session data.

module Method : sig ... end

Module with types used for matching requests.

module Query : sig ... end

Map type for working with queries from routed requests.

module Cookie : sig ... end

Less commonly used.

module MimeTypes : sig ... end

Less commonly used.

module SessionManager : sig ... end

Less commonly used.

module Router : sig ... end

Less commonly used.

module SessionConfig : sig ... end

Less commonly used.